First stab at a zib-ContactInformation implentation that is mapped on two different profiles.
The example MVPPatient shows how this can be used: there are two known slices for the two zib concepts, plus a third unknown slice for other uses. The system/use combination to designate the TelephoneNumber slice can only be used in combination with the extension -- if the extension is removed from the example, validation will fail.