Add zib BodyTemperature as a 'bare' profile directly on Observation, with the applicable HL7 Vital Signs elements added to the nl-core profile
The HL7 Vital Signs elements added are:
- The three constraints vs-1, vs-2 and vs-3
- The VSCat category

The Vital Signs IG describes two additional codes for measurement site and measurement device. The zib uses a different method to communicate this information, using the TemperatureTypeCodelist, which is added as an additional .code.coding.

In addition, the zib restricts the measurement units to just degrees Celcius, whereas in Vital Signs Fahrenheit could be used as well.