Add zib BodyHeight as a 'bare' profile directly on Observation, with the applicable HL7 Vital Signs elements added to the nl-core profile
The HL7 Vital Signs elements added are:
- The three constraints vs-1, vs-2 and vs-3
- The VSCat category
- The binding to the http://hl7.org/fhir/ValueSet/ucum-bodylength ValueSet for the value units. The zib doesn't specify which values to uses, so this is a restriction stemming from HL7.

A difference between the zib and HL7 Vital Signs is that the zib allows to specify the body position with a granularity of "standing" or "lying down", while HL7 defines an additional code for "body height measurement while lying down". However, the zib codes are qualifiers, and are thus added as an extra component. In theory, a mapping could be made from the zib concepts to additional codes, but this would probably go to far.