Add zib HeadCircumference as a 'bare' profile directly on Observation, with the applicable HL7 Vital Signs elements added to the nl-core profile
The HL7 Vital Signs elements added are:
- The three constraints vs-1, vs-2 and vs-3
- The VSCat category

In addition, there's a ValueSet bound on Observation.value to specify that just the cm and mm units may be used. This ValueSet is incompatible with the HL7 Vital Signs HeadCircumference profle, which specifies that cm and inches may be used, but not mm.
. This ValueSet is TODO in regards to naming, metadata, canonical etc.