Rethink the reversed reference for zib MedicalDevice.
The ":observation" slice is renamed to ":zibMedicalDeviceReference" and the `.short` is fixed to "zib MedicalDevice reference". The base Observation resource has been removed from this slice as to only use it for documenting references in a zib context. The derived profiles zib-VisualFunction.VisualAid and zib-HearingFunction.HearingAid no longer use this slice but define their own.