Enable the terminology server for the "loose" full repo check.
The Validator has evolved to a point where a check _with_ terminology
server is more permissive than a check without; there are now options to
respect Dutch display terms, and the Validator is more permissive about
some situations that previously resulted in errors. On the other hand,
the Validator now complains much louder on things it doesn't know off,
which is triggered by validating without a terminology server.